LPS-169954 Move all related ManagementToolbar files to a specific folder and update their imports
