+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/index.ts
@@ -28,7 +28,8 @@ export * from './components/ExpressionBuilder';
 export {FieldBase} from './components/FieldBase';
 export {Input} from './components/Input';
 export {InputLocalized} from './components/InputLocalized';
-export {ManagementToolbarSearch} from './components/ManagementToolbarSearch';
+export {ManagementToolbarSearch} from './components/ManagementToolbar/ManagementToolbarSearch';
+export {ModalEditExternalReferenceCode} from './components/ManagementToolbar/ModalEditExternalReferenceCode';
 export {RichTextLocalized} from './components/RichTextLocalized';
 export {Select} from './components/Select';
 export {CustomItem} from './components/Select/BaseSelect';
